Update apply rates action to be by benefit type
