+++ app/controllers/calculator_constants_controller.rb
@@ -3,6 +3,8 @@
 class CalculatorConstantsController < ApplicationController
   include CollectionUpdatable
 
+  before_action :set_rate_adjustment_id, only: :apply_rate_adjustments
+
   def index
     @calculator_constants = CalculatorConstant.all
     @rate_adjustments = RateAdjustment.by_chapter_number
@@ -22,10 +24,17 @@ class CalculatorConstantsController < ApplicationController
 
   # Apply rate adjustments to associated constants
   def apply_rate_adjustments
-    updated = CalculatorConstant.subject_to_rate_adjustment.each(&:apply_rate_adjustment)
+    updated = CalculatorConstant.by_rate_adjustment(@rate_adjustment_id)
+                                .each(&:apply_rate_adjustment)
     flash[:success] = {
       updated_fields: updated.pluck(:name)
     }
     redirect_to action: :index
+  end
+
+  private
+
+  def set_rate_adjustment_id
+    @rate_adjustment_id = params.require(:rate_adjustment_id)
   end
 end
