+++ app/views/calculator_constants/_constants_form.html.erb
@@ -19,5 +19,6 @@
   <div>
     <%= render partial: 'table' %>
   </div>
-  <%= render partial: "modal" %>
 <% end %>
+
+<%= render partial: "modal" %>
