+++ app/views/calculator_constants/_dropdown.html.erb
@@ -9,7 +9,7 @@
   <ul class="dropdown-menu" aria-labelledby="about-us">
     <li>
       <%= link_to "Apply Rate Adjustments", "#",
-                  data: { action: "click->rates-form#warn click->constants-form#showModal" } %>
+                  data: { action: "click->rates-form#warn click->apply-rates-modal#show" } %>
     </li>
   </ul>
 </div>
