+++ app/views/calculator_constants/_modal.html.erb
@@ -1,18 +1,22 @@
-<div data-constants-form-target="dialog" class="modal fade" role="dialog">
+<div data-apply-rates-modal-target="dialog"
+     class="modal fade"
+     role="dialog">
   <div class="modal-dialog">
     <div class="modal-content">
-      <%= form_with(url: apply_rate_adjustments_to_calculator_constants_path, method: :post) do %>
+      <!-- Form url dynamically generated by stimulus controller -->
+      <%= form_with(url: "#", method: :post, data: { apply_rates_modal_target: "form" }) do %>
         <div class="modal-header">
           <button type="button" class="close" data-dismiss="modal">&times;</button>
-          <h4 class="modal-title">Select a Benefit Type</h4>
+          <h4 class="modal-title">Apply Rate Adjustments</h4>
         </div>
         <div class="modal-body">
+          <p>Please select the benefit type for which you would like to apply rate adjustments. </p>
           <div class="form-group">
             <%= label_tag "benefit_type", "Benefit Type: "%>
-            <select name="rate_id"
+            <select name="rate_adjustment_id"
                     id="benefit_type"
-                    data-constants-form-target="select"
-                    data-action="change->constants-form#updateRateApplied">
+                    data-apply-rates-modal-target="select"
+                    data-action="change->apply-rates-modal#update">
               <%= @rate_adjustments.each do |rate_adjustment| %>
                 <option value=<%= rate_adjustment.id %>
                         data-rate=<%= rate_adjustment.rate %>
@@ -22,21 +26,23 @@
               <% end %>
             </select>
           </div>
-          <aside class="alert alert-warning" data-constants-form-target="warning">
+          <aside class="alert alert-warning" data-apply-rates-modal-target="warning">
             <p>Zero percent increase cannot be applied.</p>
-            <p>No Calculator Constants associated with benefit type.</p>
+            <p>No Calculator Constants associated with this benefit type.</p>
             <div>
-              <p>A percentage increase of <strong><span data-constants-form-target="rateApplied"></span>%</strong> will be applied to the following calculator constants:</p>
-              <ul data-constants-form-target="ul"></ul>
+              <p>A percentage increase of <strong><span data-apply-rates-modal-target="rateApplied"></span>%</strong> will be applied to the following calculator constants:</p>
+              <ul data-apply-rates-modal-target="ul"></ul>
             </div>
           </aside>
-          <p data-constants-form-target="confirm">This cannot be undone. Any unsaved changes will be lost.</p>
+          <p data-apply-rates-modal-target="confirm">This cannot be undone. Any unsaved changes will be lost.</p>
         </div>
         <div class="modal-footer">
-          <button type="button" class="btn btn-default" data-dismiss="modal">Close</button>
+          <button type="button" class="btn dashboard-btn-danger" data-dismiss="modal">Cancel</button>
+          <%= button_tag "Apply Rate Adjustments",
+                         class: "btn btn-primary",
+                         data: { apply_rates_modal_target: "applyButton" } %>
         </div>
       <% end %>
     </div>
   </div>
 </div>
-_
