+++ app/views/calculator_constants/_rates_form.html.erb
@@ -17,9 +17,9 @@
             <span class="glyphicon glyphicon-ok"></span>   
         <% end %>
         <%= button_tag type: "button",
-                      class: "btn rate-cancel",
-                      disabled: "true",
-                      data: { action: "click->rates-form#cancel", rates_form_target: "updateButton" } do %>
+                       class: "btn rate-cancel",
+                       disabled: "true",
+                       data: { action: "click->rates-form#cancel", rates_form_target: "updateButton" } do %>
         <span class="glyphicon glyphicon-remove"></span>   
         <% end %>
       </div>
