+++ app/views/calculator_constants/index.html.erb
@@ -1,5 +1,5 @@
 <div class="row row-space-6"
-     data-controller="rates-form constants-form"
+     data-controller="rates-form constants-form apply-rates-modal"
      data-action="turbo:load->rate-form#reset">
   <div class="col-md-8">
     <% if flash[:success].present? %>
