+++ config/routes.rb
@@ -48,7 +48,7 @@ Rails.application.routes.draw do
 
   resources :calculator_constants, only: [:index] do
     post :update, on: :collection
-    post :apply_rate_adjustments, on: :collection, as: :apply_rate_adjustments_to
+    post 'apply_rate_adjustments/:rate_adjustment_id', on: :collection, to: 'calculator_constants#apply_rate_adjustments'
   end
 
   resources :rate_adjustments, only: [] do
